Implement official server map indexes. (#767)
* Implement official server map indexes.
Update g_defined_map_datas with original server indexes
Update methods in map_definitions to use new indexes
Update starting zone id in EntityData
Update map id types to uint32_t as it was chopping large map id's

* Added const for default map

* Assign id to map Enviro Nightmare

+++ Projects/CoX/Common/Servers/InternalEvents.h
@@ -216,7 +216,7 @@ ONE_WAY_MESSAGE(Internal_EventTypes,ClientDisconnected)
 struct ClientMapXferData
 {
     uint64_t m_session;
-    uint8_t m_map_idx;
+    uint32_t m_map_idx;
     template<class Archive>
     void serialize(Archive &ar)
     {
